requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mysql-connector-python>=8.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pymysql>=1.1.0
mcp>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
botocore>=1.34.0